+++ client/src/components/ProfilePage/SkillGap.jsx
@@ -1,5 +1,4 @@
-import { useState, useEffect } from "react";
-import axios from "axios";
+import { useSkillMatches } from "./useSkillMatches";
 import { Bar } from "react-chartjs-2";
 import {
   Chart as ChartJs,
@@ -21,64 +20,7 @@ ChartJs.register(
 );
 
 export const SkillGap = () => {
-  const [jobs, setJobs] = useState([]);
-  const [skills, setSkills] = useState([]);
-  const [jobMatches, setJobMatches] = useState([]);
-
-  // function to compare matching and missing skills
-  const getJobMatches = (jobs, userSkillsTitles) => {
-    return jobs.map((job) => {
-      const jobSkillsTitles = job.skills.map((skill) => skill.skills_title);
-      const total = jobSkillsTitles.length;
-      const matched = jobSkillsTitles.filter((skill) =>
-        userSkillsTitles.includes(skill)
-      ).length;
-
-      const match = Math.round((matched / total) * 100);
-      const missing = 100 - match;
-
-      return {
-        title: job.jobs_title,
-        match,
-        missing,
-      };
-    });
-  };
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const jobsRes = await axios.get("http://localhost:4000/api/jobs", {
-          headers: {
-            authorization: "Bearer " + localStorage.getItem("token"),
-          },
-        });
-
-        const skillsRes = await axios.get(
-          "http://localhost:4000/api/users/skills",
-          {
-            headers: {
-              authorization: "Bearer " + localStorage.getItem("token"),
-            },
-          }
-        );
-
-        const jobs = jobsRes.data;
-        const skills = skillsRes.data.skills;
-        setJobs(jobs);
-        setSkills(skills);
-        const userSkillsTitles = skills.map((skill) => skill.title);
-        console.log("jobs:", jobs);
-        console.log("skills:", userSkillsTitles);
-        const matches = getJobMatches(jobs, userSkillsTitles);
-        console.log("Job matches:", matches);
-        setJobMatches(matches);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-    fetchData();
-  }, []);
+  const { jobMatches } = useSkillMatches();
 
   const data = {
     labels: jobMatches.map((job) => job.title),
@@ -114,7 +56,7 @@ export const SkillGap = () => {
         },
       },
       title: {
-        display: true,
+        display: false,
         text: "Skill Coverage by Job",
         color: "#004e64",
         font: {
